+++ components/App.js
@@ -560,10 +560,6 @@ class App extends Component {
             if (button === 0) {
                 if (board.get(vertex) === 0) {
                     this.makeMove(vertex)
-
-                    if (this.attachedEngineControllers.some(x => x != null)) {
-                        setTimeout(() => this.startGeneratingMoves(), setting.get('gtp.move_delay'))
-                    }
                 } else if (vertex in board.markups
                 && board.markups[vertex][0] === 'point'
                 && setting.get('edit.click_currentvertex_to_remove')) {
@@ -659,7 +655,7 @@ class App extends Component {
             if (!board.hasVertex(nextVertex)) return this.setMode('play')
 
             if (helper.vertexEquals(vertex, nextVertex)) {
-                this.makeMove(vertex, {player: 'B' in nextNode ? 1 : -1})
+                this.makeMove(vertex, {player: 'B' in nextNode ? 1 : -1, sendToEngine: false})
             } else {
                 if (board.get(vertex) !== 0
                 || this.state.blockedGuesses.some(v => helper.vertexEquals(v, vertex)))
@@ -686,7 +682,7 @@ class App extends Component {
         this.events.emit('vertexClick')
     }
 
-    makeMove(vertex, {player = null, clearUndoPoint = true} = {}) {
+    makeMove(vertex, {player = null, clearUndoPoint = true, sendToEngine = true} = {}) {
         if (!['play', 'autoplay', 'guess'].includes(this.state.mode)) {
             this.closeDrawer()
             this.setMode('play')
@@ -850,6 +846,12 @@ class App extends Component {
         // Emit event
 
         this.events.emit('moveMake', {pass, capture, suicide, ko, enterScoring})
+
+        // Send command to engine
+
+        if (sendToEngine && this.attachedEngineControllers.some(x => x != null)) {
+            setTimeout(() => this.startGeneratingMoves(), setting.get('gtp.move_delay'))
+        }
     }
 
     makeResign({player = null, setUndoPoint = true} = {}) {
@@ -861,7 +863,7 @@ class App extends Component {
         if (setUndoPoint) this.setUndoPoint('Undo Resignation')
         rootNode.RE = [`${color}+Resign`]
 
-        this.makeMove([-1, -1], {player, clearUndoPoint: false})
+        this.makeMove([-1, -1], {player, clearUndoPoint: false, sendToEngine: false})
         this.makeMainVariation(...this.state.treePosition, {setUndoPoint: false})
 
         this.events.emit('resign', {player})
@@ -1737,7 +1739,7 @@ class App extends Component {
             {
                 label: '&Clear Annotations',
                 click: () => {
-                    this.setComment({positionAnnotation: null, moveAnnotation: null})
+                    this.setComment(tree, index, {positionAnnotation: null, moveAnnotation: null})
                 }
             },
             {type: 'separator'},